#6 Création d'un menu et d'un menu pause permettant de lancer une partie, quitter la partie, quitter le jeu ou simplement mettre en pause le jeu.
