--- Assets/SavesConfig.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 42eb18f765a217d4f8f529dd786411dc
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
