--- Assets/Scripts/FollowPlayer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 29157fc30e4f2b44495a89777a57912e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
