--- Assets/Scripts/IInteractable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1fafc1e746e978247b0624c8fbe73749
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
