--- Assets/Scripts/InteractionPromptUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b3376a8cd0917cd48a4215539ca00b1c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
