+++ Assets/Scripts/Interactions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6afdf0f20001d2c4e942b3c1fe83fe7e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
