+++ Assets/Scripts/Interactions/Interactor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7724787e70f8d8d4e82e727a45732c8f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
