+++ Assets/Scripts/Mains.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cb1a0c62d171c294aac10c0fa38d8330
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
