+++ Assets/Scripts/Mains/PromptUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d695e0e5e13a81b4c887ff959efb9203
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
