+++ Assets/Scripts/Mains/Quit.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Quit : MonoBehaviour
+{
+    public void CloseProgram()
+    {
+        Application.Quit();
+    }
+}
