--- Assets/Scripts/PlayerController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ac5679881eb90174582105731ba54da2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
