--- Assets/Scripts/Question.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cb3093c1a437d8347a18e5d6652a71b4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
