--- Assets/Scripts/Quit.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Quit : MonoBehaviour
-{
-    public void CloseProgram()
-    {
-        Application.Quit();
-    }
-}
