--- Assets/Scripts/Quit.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1b41a756baa547c4285a7b2bf02a5f0d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
