--- Assets/Scripts/SceneLoader.cs
@@ -1,27 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class SceneLoader : MonoBehaviour
-{
-
-    private string _scenes;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-    //This is gonna open a scene.
-    public void Load(string sceneName)
-    {
-        SceneManager.LoadScene(sceneName);
-    }
-}
