--- Assets/Scripts/SceneLoader.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4e171e682e4c77b48861e3116a67f51f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
