--- Assets/Scripts/UnityEventInteractor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5daa0df52f2210046a7cacc467fa66b3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
